// src/app/core/models/organization/business_profile/organization-chart-image.model.ts
export interface OrganizationChartImageResponse{
    created_at: string;
    created_by: number;
    created_by_first_name: string;
    created_by_last_name: string;
    ext: string;
    id: number;
    organization_id: number;
    organization_title: string;
    size: number;
    status: string;
    status_id: number;
    status_label: string;
    thumbnail_url: string;
    title: string;
    token: string;
    url: string;
}